feature: rename exec util to process util and init exec and spawn utility
