+++ src/providers/source-control.ts
@@ -5,7 +5,7 @@ import { getStylesheetURI } from '../utils/stylesheet'
 import { getScriptURI } from '../utils/script'
 import { isWorkspaceHasGit } from '../utils/workspace'
 import { renderIntroUI, renderCommitFormUI, renderOpenFolderUI } from '../ui'
-import exec from '../utils/exec'
+import { spawn } from '../utils/process'
 
 class SourceControlProvider implements vscode.WebviewViewProvider {
   public static readonly type = 'commitizen-code.source-control'
@@ -46,7 +46,7 @@ class SourceControlProvider implements vscode.WebviewViewProvider {
             `Commitizen Code: Commit message is: ${message.data.commit} `
           )
 
-          exec({ exec: 'git', cmd: 'status' })
+          spawn({ exec: 'git', cmd: 'status' })
 
           break
         }
